`@remotion/webcodecs`: `convertMedia()` for converting to webM
