+++ packages/webcodecs/src/index.ts
@@ -1,4 +1,5 @@
 export {createAudioDecoder} from './audio-decoder';
 export {createAudioEncoder} from './audio-encoder';
+export {ConvertMediaState, convertMedia} from './convert-media';
 export {createVideoDecoder} from './video-decoder';
 export {createVideoEncoder} from './video-encoder';
